Add the ability to reauth with a different mechanism
This commit adds an ".auth_with(opts)" method that allows clients to
create a new connection with an alternative auth mechanism if required.

For example, let's say a client needs to be able to hit the root
collection with the HMAC auth strategy since they do not have a user
token with which to do so. Then can simply init with the HMAC id and
secret.

However, if they need to run requests with user tokens, they are able to
do so by establishing a new connection with `TeamSnap.auth_with(:token
=> "token")` which will create a new Faraday client with the appropriate
auth credentials. Any subsequent requests will then use the most recent
auth mechanism provided.