Alternative Testing Method
Throwing this out as an alternative. Since the init method itself
needn't actually init anything to describe it's behavior, what if we
simple stubbed the http requests? This allows the tests to be super
fast, not rely on a cassette, and exercise exactly what we're trying to
describe in this method without an need to also be testing the auth.